feat: 1. Cactus Auction UI/UX 기능 구현
질문
MainFragment 에서 CactusBasketRecyclerViewAdapter 는 fragment 전환해도 1번만 init이 되는데
AuctionFragment 에서는 CactusAuctionBasketRecyclerViewAdapter 가 fragment 전환 마다 init이 호출 된다.

+++ app/src/main/java/com/cjwgit/jejucactusreceipt/ui/BindingAdapter.kt
@@ -30,7 +30,7 @@ fun isDoubleClick(): Boolean {
 
 // region Int Text를 ###,### 포맷팅 하는 바인딩 아답터
 @BindingAdapter("android:formattedText")
-fun setFormattedText(view: TextView, value: Int) {
+fun setFormattedText(view: TextView, value: Long) {
     val formattedValue = DecimalFormat("###,###").format(value)
     view.text = formattedValue
 }
